+++ src/Controller/GameBuilder.h
@@ -11,12 +11,12 @@
 class GameBuilder {
 public:
 	GameBuilder();
-	void withShape(Shape<int> * shape);
+	void withShape(Shape<double> * shape);
 	void withButton(Button * button);
 	Game * build(const char * filename);
 
 private:
 	Menu * menu;
-	std::vector<Shape<int> *> pieces;
+	std::vector<Shape<double> *> pieces;
 };
